Add --tarball option to create a tar of the disk
The configuration of debootstrap is useful even if a disk image is
not created, and allows the same configuration to be used for targets
that don't accept a raw disk image.

It's possible to use both --image and --tarball, in which case
it behaves the same as --image, but also creates a tarball of the
contents of the disk image.

If --tarball is specified, but --image isn't then, instead of creating
a disk image, it will debootstrap into a temporary directory.